backend/requirements.txt
Django>=5.0,<5.3
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
drf-spectacular>=0.27.2
psycopg[binary]>=3.2.0
dj-database-url>=2.2.0
whitenoise>=6.7.0
gunicorn>=22.0.0
python-dotenv>=1.0.1
django-filter>=24.2
